refactor: split OpenAI responses API module
